refactor simple factory pkg
