+++ src/main/java/designpattern/simplefactory/Calculator.java
@@ -1,8 +1,10 @@
-package main.java.designpattern;
+package main.java.designpattern.simplefactory;
 
 public class Calculator{
   public static void main(String[] args){
+    /*运算符*/
     String operation = args[1];
+    /*两个运算数字*/
     double n1 = Double.parseDouble(args[0]);
     double n2 = Double.parseDouble(args[2]);
