+++ src/main/java/designpattern/simplefactory/Operator.java
@@ -1,4 +1,4 @@
-package main.java.designpattern;
+package main.java.designpattern.simplefactory;
 
 public abstract class Operator{
   public abstract double cal(double n1, double n2);
