+++ src/main/java/designpattern/simplefactory/OperatorFactory.java
@@ -1,4 +1,4 @@
-package main.java.designpattern;
+package main.java.designpattern.simplefactory;
 
 public class OperatorFactory {
   public static Operator createOperator(String operation){
